Feature/#39 schedule controller service impl

+++ src/main/java/show/schedulemanagement/service/schedule/ScheduleServiceImpl.java
@@ -1,11 +1,17 @@
 package show.schedulemanagement.service.schedule;
 
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import show.schedulemanagement.domain.member.Member;
 import show.schedulemanagement.domain.schedule.Schedule;
+import show.schedulemanagement.domain.schedule.ScheduleAble;
 import show.schedulemanagement.repository.schedule.ScheduleRepository;
+import show.schedulemanagement.service.MemberService;
 
 @Service
 @RequiredArgsConstructor
@@ -14,22 +20,41 @@ import show.schedulemanagement.repository.schedule.ScheduleRepository;
 public class ScheduleServiceImpl implements ScheduleService{
 
     private final ScheduleRepository scheduleRepository;
+    private final MemberService memberService;
 
     @Override
     @Transactional
     public void save(Schedule schedule) {
         scheduleRepository.save(schedule);
-        //파라미터로 부터 -> 일반일정 세부사항 여러개를 만들고 -> 이 각 세부사항의 범위
     }
 
     @Override
     @Transactional
-    public void deleteById(Long id) {
-        scheduleRepository.deleteById(id);
+    public void deleteById(Long id) throws Exception {
+        Member member = memberService.getAuthenticatedMember();
+        Schedule schedule = findById(id);
+        String createdBy = schedule.getCreatedBy();
+        if(member.getEmail().equals(createdBy)){
+            scheduleRepository.deleteById(id);
+        }
+        throw new Exception("작성자만 삭제할 수 있습니다.");
     }
 
     @Override
     public Schedule findById(Long id) {
         return scheduleRepository.findById(id).orElseThrow(() -> new RuntimeException("Schedule Not Found"));
+    }
+
+    @Override
+    public Schedule getConflictSchedule(List<ScheduleAble> newSchedule, Member member, LocalDateTime start, LocalDateTime end) {
+        List<Schedule> allSchedule = scheduleRepository.findAllBetweenStartAndEnd(member, start, end);
+        for (Schedule schedule : allSchedule) {
+            for (ScheduleAble scheduleAble : newSchedule) {
+                if(schedule.isConflict(scheduleAble)){
+                    return schedule;
+                }
+            }
+        }
+        return null;
     }
 }
